refactor(gui): split ThemeProvider out of useTheme hook
Move ThemeProvider component from hooks/useTheme.tsx into its own
file at components/ThemeProvider.tsx. The hook file now exports only
the hook, types, and context — resolving the react-refresh lint
error that occurs when a file exports both components and functions.
