+++ frontend/src/hooks/useTheme.tsx
@@ -1,17 +1,9 @@
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
-import { send, subscribe } from "./useIPC";
+import { createContext, useContext } from "react";
 
 export type ThemeMode = "light" | "dark" | "system";
 export type ResolvedTheme = "light" | "dark";
 
-type ThemeContextValue = {
+export type ThemeContextValue = {
   /** User's stored preference — may be "system". */
   mode: ThemeMode;
   /** Concrete theme currently applied ("light" | "dark"). */
@@ -20,65 +12,13 @@ type ThemeContextValue = {
   setMode: (m: ThemeMode) => void;
 };
 
-const ThemeContext = createContext<ThemeContextValue | null>(null);
+export const ThemeContext = createContext<ThemeContextValue | null>(null);
 
-function detectSystem(): ResolvedTheme {
+export function detectSystem(): ResolvedTheme {
   if (typeof window === "undefined" || !window.matchMedia) return "dark";
   return window.matchMedia("(prefers-color-scheme: light)").matches
     ? "light"
     : "dark";
-}
-
-function normalize(raw: unknown): ThemeMode {
-  return raw === "light" || raw === "dark" ? raw : "system";
-}
-
-export function ThemeProvider({ children }: { children: React.ReactNode }) {
-  const [mode, setModeState] = useState<ThemeMode>("system");
-  const [systemTheme, setSystemTheme] = useState<ResolvedTheme>(() =>
-    detectSystem(),
-  );
-
-  useEffect(() => {
-    const unsub = subscribe((msg) => {
-      if (msg.type === "theme") {
-        setModeState(normalize(msg.mode));
-      }
-    });
-    send({ type: "theme_get" });
-    return unsub;
-  }, []);
-
-  useEffect(() => {
-    if (typeof window === "undefined" || !window.matchMedia) return;
-    const mql = window.matchMedia("(prefers-color-scheme: light)");
-    const onChange = () => setSystemTheme(mql.matches ? "light" : "dark");
-    // Safari < 14 still uses addListener; guard for both.
-    if (mql.addEventListener) mql.addEventListener("change", onChange);
-    else mql.addListener(onChange);
-    return () => {
-      if (mql.removeEventListener) mql.removeEventListener("change", onChange);
-      else mql.removeListener(onChange);
-    };
-  }, []);
-
-  const resolved: ResolvedTheme = mode === "system" ? systemTheme : mode;
-
-  useEffect(() => {
-    document.documentElement.setAttribute("data-theme", resolved);
-  }, [resolved]);
-
-  const setMode = useCallback((next: ThemeMode) => {
-    setModeState(next);
-    send({ type: "theme_set", mode: next });
-  }, []);
-
-  const value = useMemo(
-    () => ({ mode, resolved, setMode }),
-    [mode, resolved, setMode],
-  );
-
-  return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>;
 }
 
 export function useTheme(): ThemeContextValue {
